[BE] rename: constant 폴더 추가
