+++ BE/newsainturtle/src/main/java/com/newsainturtle/common/exception/JwtException.java
@@ -3,7 +3,7 @@ package com.newsainturtle.common.exception;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ResponseStatus;
 
-import static com.newsainturtle.common.CommonConstant.*;
+import static com.newsainturtle.common.constant.CommonConstant.*;
 
 @ResponseStatus(HttpStatus.FORBIDDEN)
 public class JwtException extends BadRequestException {
